Some vars, arguments of lists using size_t on x86-64 mode are 'too much'!

+++ src/include/t50_modules.h
@@ -100,12 +100,12 @@ typedef struct
  * The modules table is global through all the code.
  */
 extern modules_table_t mod_table[]; // Must be extern here!
-extern const size_t number_of_modules;
-extern size_t indices[];
+extern const uint32_t number_of_modules;
+extern uint32_t indices[];
 
 int    *get_module_valid_options_list(int);
 void    build_indices(void);
-size_t  get_index(struct config_options *);
+uint32_t get_index(struct config_options *);
 
 /* Modules functions prototypes. */
 void icmp  (const struct config_options *const __restrict__, size_t *);
